requirements.txt
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.2
MouseInfo==0.1.3
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
six==1.16.0
tzdata==2024.1
